oficios-automation/funcoes/W5_archiving/requirements.txt
google-cloud-firestore>=2.14.0
google-cloud-storage>=2.10.0
google-cloud-pubsub>=2.18.0
pydantic>=2.5.0
python-dateutil>=2.8.2





